Fix github actions jobs
This patch reworks the github actions declaration file to have less
duplication, easier to maintain, and easier to add new configurations
to the build matrix.

You can see the result on my personal fork:
https://github.com/violinyanev/ramses-logic/pull/7

It builds 4 things now - Windows, Ubuntu with GCC, Ubuntu with Clang,
and Ubuntu with Clang-tidy (only lint, no build). This has also the
side effect that the jobs run faster, since the clang build and clang
tidy are split in two jobs now (they don't need to run on the same job).

The matrix configuration can be read like this:
* Run on Ubuntu and Windows
* Use the default compiler unless specified differently
* Run an extra Ubuntu job which uses Clang instead of default compiler (GCC)
* Run an extra Ubuntu job with Clang-tidy